requirements.txt
openai==1.51.2
fastapi==0.115.2
pandas==2.2.3
pydantic==2.9.2
uvicorn==0.32.0
python-dotenv==1.0.1 
